requirements.txt
llama-cpp-python>=0.2.0 